requirements.txt
Django==1.10.5
python-social-auth==0.3.3
django-social-auth==0.7.28
social-auth-app-django==0.1.0
pytz==2016.10
django-admin-sortable==2.0.21
easy-thumbnails==2.3
django-image-cropping==1.0.3
